Adds options databases and reasons
This commit adds two new arguments to the prisma-functions: 'databases' and 'reasons'. They both accept a list with two vectors of identical lenghts.
'databases' adds names of databases with the number of found articles in each database to the label 'found'.
'reasons' adds reasons with the number of articles excluded with that reason to the label 'full_text_exclusions'.
Checks are performed to ensure the arguments are given in the right format.
A warning is returned if the sum of the numerical vector doesn't match the stated 'found' oder 'full_text_exclusion' number.